Add ConfirmationInstructionsLive and tests
