+++ priv/templates/phx.gen.auth/confirmation_live_test.exs
@@ -78,52 +78,4 @@ defmodule <%= inspect context.web_module %>.<%= inspect Module.concat(schema.web
       refute <%= inspect context.alias %>.get_<%= schema.singular %>!(<%= schema.singular %>.id).confirmed_at
     end
   end
-
-  describe "Resend confirmation" do
-    test "renders the resend confirmation page", %{conn: conn} do
-      {:ok, _lv, html} = live(conn, Routes.<%= schema.route_helper %>_confirmation_path(conn, :new))
-      assert html =~ "<h1>Resend confirmation instructions</h1>"
-    end
-
-    test "sends a new confirmation token", %{conn: conn, <%= schema.singular %>: <%= schema.singular %>} do
-      {:ok, lv, _html} = live(conn, Routes.<%= schema.route_helper %>_confirmation_path(conn, :new))
-
-      {:ok, conn} =
-        lv
-        |> form("#resend_confirmation_form", <%= schema.singular %>: %{email: <%= schema.singular %>.email})
-        |> render_submit()
-        |> follow_redirect(conn, "/")
-
-      assert get_flash(conn, :info) =~ "If your email is in our system"
-      assert Repo.get_by!(<%= inspect context.alias %>.<%= inspect schema.alias %>Token, <%= schema.singular %>_id: <%= schema.singular %>.id).context == "confirm"
-    end
-
-    test "does not send confirmation token if <%= schema.singular %> is confirmed", %{conn: conn, <%= schema.singular %>: <%= schema.singular %>} do
-      Repo.update!(<%= inspect context.alias %>.<%= inspect schema.alias %>.confirm_changeset(<%= schema.singular %>))
-
-      {:ok, lv, _html} = live(conn, Routes.<%= schema.route_helper %>_confirmation_path(conn, :new))
-
-      {:ok, conn} =
-        lv
-        |> form("#resend_confirmation_form", <%= schema.singular %>: %{email: <%= schema.singular %>.email})
-        |> render_submit()
-        |> follow_redirect(conn, "/")
-
-      assert get_flash(conn, :info) =~ "If your email is in our system"
-      refute Repo.get_by(<%= inspect context.alias %>.<%= inspect schema.alias %>Token, <%= schema.singular %>_id: <%= schema.singular %>.id)
-    end
-
-    test "does not send confirmation token if email is invalid", %{conn: conn} do
-      {:ok, lv, _html} = live(conn, Routes.<%= schema.route_helper %>_confirmation_path(conn, :new))
-
-      {:ok, conn} =
-        lv
-        |> form("#resend_confirmation_form", <%= schema.singular %>: %{email: "unknown@example.com"})
-        |> render_submit()
-        |> follow_redirect(conn, "/")
-
-      assert get_flash(conn, :info) =~ "If your email is in our system"
-      assert Repo.all(<%= inspect context.alias %>.<%= inspect schema.alias %>Token) == []
-    end
-  end
 end
